LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/grid-template-columns-rows-expected.txt


PASS Can set 'grid-template-columns' to CSS-wide keywords: initial
PASS Can set 'grid-template-columns' to CSS-wide keywords: inherit
PASS Can set 'grid-template-columns' to CSS-wide keywords: unset
PASS Can set 'grid-template-columns' to CSS-wide keywords: revert
PASS Can set 'grid-template-columns' to var() references:  var(--A)
PASS Can set 'grid-template-columns' to the 'none' keyword: none
FAIL Setting 'grid-template-columns' to a length throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
FAIL Setting 'grid-template-columns' to a percent throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'grid-template-columns' to a time throws TypeError
PASS Setting 'grid-template-columns' to an angle throws TypeError
FAIL Setting 'grid-template-columns' to a flexible length throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
FAIL Setting 'grid-template-columns' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'grid-template-columns' to a URL throws TypeError
PASS Setting 'grid-template-columns' to a transform throws TypeError
FAIL 'grid-template-columns' does not supported '[linename1] 100px [linename2 linename3]' assert_class_string: Unsupported value must be a CSSStyleValue and not one of its subclasses expected "[object CSSStyleValue]" but got "[object CSSKeywordValue]"
FAIL 'grid-template-columns' does not supported '200px repeat(auto-fill, 100px) 300px' assert_class_string: Unsupported value must be a CSSStyleValue and not one of its subclasses expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
PASS Can set 'grid-template-rows' to CSS-wide keywords: initial
PASS Can set 'grid-template-rows' to CSS-wide keywords: inherit
PASS Can set 'grid-template-rows' to CSS-wide keywords: unset
PASS Can set 'grid-template-rows' to CSS-wide keywords: revert
PASS Can set 'grid-template-rows' to var() references:  var(--A)
PASS Can set 'grid-template-rows' to the 'none' keyword: none
FAIL Setting 'grid-template-rows' to a length throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
FAIL Setting 'grid-template-rows' to a percent throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'grid-template-rows' to a time throws TypeError
PASS Setting 'grid-template-rows' to an angle throws TypeError
FAIL Setting 'grid-template-rows' to a flexible length throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
FAIL Setting 'grid-template-rows' to a number throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'grid-template-rows' to a URL throws TypeError
PASS Setting 'grid-template-rows' to a transform throws TypeError
FAIL 'grid-template-rows' does not supported '[linename1] 100px [linename2 linename3]' assert_class_string: Unsupported value must be a CSSStyleValue and not one of its subclasses expected "[object CSSStyleValue]" but got "[object CSSKeywordValue]"
FAIL 'grid-template-rows' does not supported '200px repeat(auto-fill, 100px) 300px' assert_class_string: Unsupported value must be a CSSStyleValue and not one of its subclasses expected "[object CSSStyleValue]" but got "[object CSSUnitValue]"
